start moving methods out of gtfsrt tasks
